Fixing widget CSS/JS imports for Django 1.8
In Django 1.8, trailing slashes on `STATIC_URL` is enforced, so the import URL for the CSS and JS doesn't need to add it. In fact, with it, I get breakage on the admin page, with a bunch of errors of the form:

```
GET http://127.0.0.1:8000/static//paintstore/jquery_1.7.2.js 
```
for each static resource paintstore requires.

This fixes it, at least on Django 1.8.